fix: fetch team details using events collection
also changed api route location
